+++ src/routes/nest-visits/index.js
@@ -1,0 +1,13 @@
+const batchCreateVisits = require('./batch-create-visits.js')
+
+module.exports = function (fastify, opts, next) {
+  fastify.post(
+    '/nest-visits/batch',
+    { schema: batchCreateVisits.schema },
+    batchCreateVisits.handler
+  )
+
+  next()
+}
+
+module.exports.prefixOverride = ''