+++ src/routes/nests/index.js
@@ -11,10 +11,18 @@ const {
   schema: createNestVisitSchema,
   handler: createNestVisitHandler
 } = require('./create-nest-visit.js')
+const batchCreateNests = require('./batch-create-nests.js')
 
 module.exports = function (fastify, opts, next) {
   fastify.get('/nests', { schema: getAllSchema }, getAllHandler)
   fastify.post('/nests', { schema: postNestSchema }, postNestHandler)
+
+  fastify.post(
+    '/nests/batch',
+    { schema: batchCreateNests.schema },
+    batchCreateNests.handler
+  )
+
   fastify.put('/nests/:nestId', { schema: editNestSchema }, editNestHandler)
   fastify.post(
     '/nests/:nestId/visits',